Move runtime_env_disable_URI_cache fixture
