+++ python/ray/tests/test_runtime_env_working_dir_3.py
@@ -6,6 +6,7 @@ import tempfile
 
 import pytest
 from pytest_lazyfixture import lazy_fixture
+from unittest import mock
 
 import ray
 from ray.exceptions import GetTimeoutError
@@ -21,6 +22,21 @@ from ray._private.utils import get_directory_size_bytes
 # Calling `test_module.one()` should return `2`.
 # If you find that confusing, take it up with @jiaodong...
 S3_PACKAGE_URI = "s3://runtime-env-test/test_runtime_env.zip"
+
+
+# Set scope to "class" to force this to run before start_cluster, whose scope
+# is "function".  We need these env vars to be set before Ray is started.
+@pytest.fixture(scope="class")
+def runtime_env_disable_URI_cache():
+    with mock.patch.dict(
+        os.environ,
+        {
+            "RAY_RUNTIME_ENV_WORKING_DIR_CACHE_SIZE_GB": "0",
+            "RAY_RUNTIME_ENV_PY_MODULES_CACHE_SIZE_GB": "0",
+        },
+    ):
+        print("URI caching disabled (cache size set to 0).")
+        yield
 
 
 @pytest.mark.skipif(
